Update patch set 4
Patch Set 4:

> Patch Set 1: Code-Review-1
> 
> Could you run code analyze from IntelliJ (Analize->Inspect Code), next execute https://plugins.jetbrains.com/plugin/7973-sonarlint verification then fix all reported violation?
> 
> Additionally pls remove cyclic dependency between classes from org.onap.aaf.certservice.cmpv2client.api and org.onap.aaf.certservice.cmpv2client.impl packages.
> 
> To check cyclic dependency pls use IntelliJ option: Analyze -> Analyze Cyclic Dependency...
> 
> P.s. Today I will try to do code review.


Fixed the cyclic dependencies and the majority of the violations listed by Intellij's analysis. The majority of Violations I left in were concerning access modifier for the classes, and for many of the classes I want them to remain public etc. so that the methods can be shared and reused if necessary. Feel free to ask about any of the violations you might see on the code though - I can justify why I left them in for you.

Patch-set: 4