Update patch set 5
Patch Set 5:

Actually , your job is passing using the new docs that we are planning to migrate to: https://jenkins.onap.org/job/rtdv3-global-verify-master/198/ : SUCCESS

Since we are still not deprecating the submodules, I think we will need to merge the change equivalent to this that you guys had initially to add the submodule :( Sorry i confused you guys. 
Do you still have that change?

We need to merge both, the submodule change and then re-verify this one and merge.

Patch-set: 5